feat: move shared helpers to separate modules
