+++ src/api/content/upload.ts
@@ -18,7 +18,6 @@ export default async function handler(req: Request, res: Response) {
     const [fields, files] = (await form.parse(req)) as [Fields, Files];
     const blockId = parseInt(fields.blockId?.toString() ?? '', 10);
     const durationMinutes = fields.durationMinutes ? parseInt(fields.durationMinutes.toString(), 10) : null;
-    const file = files.file as File | File[] | undefined;
     const file = (files as Files & { file?: File | File[] }).file;
