+++ src/components/ui/sonner-toast.ts
@@ -1,0 +1,1 @@
+export { toast } from "sonner"
