+++ src/lib/auth.tsx
@@ -1,22 +1,5 @@
-import React, { createContext, useContext, useState, useEffect } from 'react';
-
-export type UserRole = 'ADMIN' | 'ASSESSOR' | 'VIEWER' | 'ENTERPRISE';
-
-export interface User {
-  id: string;
-  email: string;
-  role: UserRole;
-  name: string;
-}
-
-interface AuthContextType {
-  user: User | null;
-  login: (email: string, password: string) => Promise<boolean>;
-  logout: () => void;
-  isLoading: boolean;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
+import React, { useState, useEffect } from 'react';
+import { AuthContext, type User, type UserRole } from './use-auth';
 
 // Mock users for demo
 const mockUsers: User[] = [
@@ -87,12 +70,4 @@ export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children
       {children}
     </AuthContext.Provider>
   );
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
 };
