+++ src/pages/__tests__/Dashboard.test.tsx
@@ -3,7 +3,7 @@ import { MemoryRouter } from 'react-router-dom';
 import { vi } from 'vitest';
 
 import Dashboard from '@/pages/Dashboard';
-import { useAuth, type UserRole } from '@/lib/auth';
+import { useAuth, type UserRole } from '@/lib/use-auth';
 
 // Mock hooks and contexts used within Dashboard
 vi.mock('@/hooks/use-toast', () => ({
@@ -15,7 +15,7 @@ vi.mock('@/hooks/use-enterprise-branding', () => ({
   default: () => {},
 }));
 
-vi.mock('@/lib/auth', () => ({
+vi.mock('@/lib/use-auth', () => ({
   useAuth: vi.fn(),
 }));
